To align with best pytest practices, move common fixtures to conftest.py rather than importing them from other test files
