+++ tests/test_tarfiles_unit.py
@@ -25,7 +25,6 @@ import tarfiles
 import get_parser
 
 from test_unit import TestUnit
-from test_creds_unit import needs_credentials
 
 
 class TestTarfilesUnit:
